refactor(local-service): split execution tool helpers
